mqtt_site/requirements.txt
Django==4.1
django-htmx==1.12.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
paho-mqtt==1.6.1
